cancel size rounding

+++ modules/atomic-widgets/prop-types/size-prop-type.php
@@ -29,7 +29,7 @@ class Size_Prop_Type extends Plain_Prop_Type {
 
 	protected function sanitize_value( $value ) {
 		return [
-			'size' => (int) $value['size'],
+			'size' => (float) $value['size'],
 			'unit' => sanitize_text_field( $value['unit'] ),
 		];
 	}
